Fixed a bug in command response handling.
It appears that, when using MPSSE_BITMODE with <8 bits, the low
bits of the shift register can contain old data.  In this case,
we were reading 3 bits -- which come back in the MSBs of the shift
register -- but also getting 1s in bits 1 and 2.  Looks like they
were left over from a previous transaction.