Don't send SWD data phase on WAIT/FAULT.
The Cortex-M0 rarely generates WAIT.  The M3, on the other
hand, is all sophistimacated and stuff and frequently generates
WAIT responses when accessing memory.  With overrun detection
disabled (the current default) it interprets a data phase
as a framing error and goes off into the weeds.

Somewhat alarmingly, these framing errors appear to survive a
chip reset.  I'm currently having to power-cycle to recover.

Anyway, this commit eliminates the framing errors I'm aware of.