A grab-bag of small fixes to swddude for ARMv7-M.
Fixed Flash chunking algorithm for programs >256 bytes.

Enabled faults for better reporting of crashes in IAP firmware.

Explicitly allocated stack, instead of assuming that the stack
pointer is valid.  This is important because the SP is often near
the top of RAM when halted -- an area used for scratch by the IAP.