Added connect-during-reset and vector catch support.
This makes the initial programming environment a lot more
predictable, and avoids taking a fault in buggy or corrupted
firmware.  It's also what the STM32 wants us to do every time.

One caveat: on the LPCs, I believe this prevents the bootloader
from executing before programming.  It doesn't seem to be a huge
problem, but it may prevent some state from being initialized.
The right thing to do is have the target in control of the
reset-halt operation.