Target: fixed private method defn order.
The function definitions in target.cpp are literally in the order I
wrote them, top to bottom.  This is kind of confusing.  In this
change I've fixed the *private* methods to match order of declaration
in the header.  While I was moving things around and generating an
ugly diff anyway, I took the opportunity to reformat the private
methods to match our current conventions.