Target: introduced rptr<T>.
Previously, addresses in the target's memory were represented as
simple uint32_t values.  This had some disadvantages: it made the
purpose less clear; it made pointer arithmetic manual; it didn't
distinguish between types of pointers or const-nexx.

With this change, target addresses are stored using "remote
pointers" or rptrs.

Because of the extensive diff this produced, I took the opportunity
to do more cleanups.