Target: fixed ARMv7-Mism that crept into reset code.
Setting SYSRESETREQ works on both v6-M and v7-M, though it's a bit
heavy-handed on v7-M and wanders dangerously close to some errata
on early versions of the Cortex-M3.

The right thing to do here is to split Target's logic up into
separate reset strategies for each architecture.