FTDI: Move towards supporting more FTDI based programmers
This is the first step towards a more robust programmer enumeration
strategy.  We now use libusb to locate the ftdi chip we will use.
This still just looks up the part by VID:PID which is problematic,
but libusb provides the functionality required to traverse the bus
looking for a specific device based on other parameters.

This commit also adds vid, pid, and interface command line parameters
for controlling which device is selected and which interface port of
the selected device to use.