Programmer: Add support for BusBlaster
This change includes new support for selecting a programmer.  The
default programmer is the um232h evaluation board from FTDI since
that was the first board that worked.  Once a programmer has been
selected the vid, pid, and interface can all be overridden from the
command line.

The BusBlaster must have been programmed with the KT-link RTL.
Consiquently, this version should now support the KT-link as well,
but that has not been tested.