Target: memory accesses use banked registers.
This lets us access 16 bytes of memory without changing the TAR
(Transfer Address Register).  It amplifies the benefits of write
coalescing on TAR: things like core register accesses require a
sequence of memory accesses to adjacent registers, which are now
single SWD transactions.

swddump is 23.5% faster with this change.

swddude itself is 9% faster on LPC1343 (its runtime is dominated
by calling and waiting on the IAP ROM).