MPSSE: fixed double-free of _handle.
If we can't kick the kernel driver off our device, the cleanup code
was closing (=> freeing) the libusb handle, but not zeroing the
pointer.  The destructor would then misinterpret it as a valid handle
and attempt to free it again.

While this bug theoretically affects all platforms, it only showed
up on Mac OS X because (1) libftdi can't evict the kernel driver and
(2) malloc error checking is on by default.