swddude: fixed two errors for programs > 256 bytes.
One caused a sector number off-by-one error by rounding down rather
than up.

The other was a simple min/max confusion that caused the RAM buffer
for code to stomp on the working area.  Due to the order in which
we write the buffers, this didn't actually cause problems, but for
very large programs it could have caused faults.